Extract the common domain picker logic into a base class with specific use-case implementations
